Removed Vault Simplified permissions to signurls.admin Updated to 1.15.1
