--- SignURLs/.gitignore
@@ -1,5 +1,0 @@
-bin
-.metadata
-.classpath
-.project
-.settings
