+++ pom.xml
@@ -1,0 +1,35 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>net.willhastings</groupId>
+  <artifactId>SignURLs</artifactId>
+  <version>0.0.1-SNAPSHOT</version>
+  <name>SignURLs</name>
+  <description>Add URLs to Signs</description>
+  <build>
+    <sourceDirectory>src</sourceDirectory>
+    <plugins>
+      <plugin>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.8.0</version>
+        <configuration>
+          <source>1.8</source>
+          <target>1.8</target>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+  <repositories>
+    <repository>
+      <id>maven-snapshots</id>
+      <url>https://papermc.io/repo/repository/maven-snapshots/</url>
+    </repository>    
+  </repositories>
+  <dependencies>
+    <dependency>
+	  <groupId>com.destroystokyo.paper</groupId>
+	  <artifactId>paper-api</artifactId>
+	  <version>1.14.4-R0.1-SNAPSHOT</version>
+	  <scope>provided</scope>
+	</dependency>
+  </dependencies>
+</project>
